6 sept

+++ GFG/Stack/MaxRectAreaHistogram.java
@@ -3,11 +3,11 @@ package GFG.Stack;
 import java.util.Stack;
 
 public class MaxRectAreaHistogram {
-    public static long getMaxArea(long arr[], long n) {
+    public static int getMaxArea(int arr[], int n) {
         Stack<Integer> s = new Stack<>();
-        long res = 0;
+        int res = 0;
         int tp;
-        long curr;
+        int curr;
         for (int i = 0; i < n; i++) {
             while (s.isEmpty() == false && arr[s.peek()] >= arr[i]) {
                 tp = s.peek();
@@ -29,7 +29,7 @@ public class MaxRectAreaHistogram {
 
     public static void main(String[] args) {
         int N = 7;
-        long arr[] = { 6, 2, 5, 4, 5, 1, 6 };
+        int arr[] = { 6, 2, 5, 4, 5, 1, 6 };
         System.out.println(getMaxArea(arr, N));
     }
 }
